♻️  Split tokio stream into a separate file
